requirements/requirements_v3.txt
flask_script
flask-mysql
flask_bcrypt
sqlalchemy
flask_paginate
opencv-python
pillow
scipy
nibabel
mxnet
imageio
mysql-connector-python
